+++ src/pages/Bloglist.jsx
@@ -11,16 +11,26 @@ const Bloglist = () => {
   useEffect(() => {
     const fetchCategories = async () => {
       try {
-        const token = localStorage.getItem("access_token");
-        const response = await axios.get(
-          "http://localhost:5000/api/blogcategory/",
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          }
-        );
-        setCategories(response.data);
+       const token = JSON.parse(localStorage.getItem("access_token"));
+       if (
+         token &&
+         token.expirationDate &&
+         new Date() > new Date(token.expirationDate)
+       ) {
+         // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+         alert("Token is expired, please login again.");
+       } else {
+         // Token còn hiệu lực, tiếp tục sử dụng
+         const response = await axios.get(
+           "http://localhost:5000/api/blogcategory/",
+           {
+             headers: {
+               Authorization: `Bearer ${token}`,
+             },
+           }
+         );
+         setCategories(response.data);
+       }
       } catch (error) {
         console.error(error);
       }
@@ -35,13 +45,23 @@ const Bloglist = () => {
   });
   const fetchData = async () => {
     try {
-      const token = localStorage.getItem("access_token");
-      const response = await axios.get("http://localhost:5000/api/blog/", {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      setData(response.data);
+     const token = JSON.parse(localStorage.getItem("access_token"));
+     if (
+       token &&
+       token.expirationDate &&
+       new Date() > new Date(token.expirationDate)
+     ) {
+       // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+       alert("Token is expired, please login again.");
+     } else {
+       // Token còn hiệu lực, tiếp tục sử dụng
+       const response = await axios.get("http://localhost:5000/api/blog/", {
+         headers: {
+           Authorization: `Bearer ${token}`,
+         },
+       });
+       setData(response.data);
+     }
     } catch (error) {
       throw new Error(error);
     }
@@ -52,13 +72,23 @@ const Bloglist = () => {
   const onDeleteBlog = async (id, e) => {
     if (window.confirm("Bạn có chắc chắn muốn xóa dữ liệu này không?")) {
       try {
-        const token = localStorage.getItem("access_token");
-        await axios.delete(`http://localhost:5000/api/blog/${id}`, {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        fetchData();
+        const token = JSON.parse(localStorage.getItem("access_token"));
+        if (
+          token &&
+          token.expirationDate &&
+          new Date() > new Date(token.expirationDate)
+        ) {
+          // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+          alert("Token is expired, please login again.");
+        } else {
+          // Token còn hiệu lực, tiếp tục sử dụng
+          await axios.delete(`http://localhost:5000/api/blog/${id}`, {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          });
+          fetchData();
+        }
       } catch (error) {
         throw new Error(error);
       }
@@ -77,23 +107,33 @@ const Bloglist = () => {
   const handleUpdateBlog = async (e) => {
     const { id, title, category, description } = updateData;
     try {
-      const token = localStorage.getItem("access_token");
-      await axios.put(
-        `http://localhost:5000/api/blog/${id}`,
-        {
-          title,
-          category,
-          description,
-        },
-        {
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      handleCloseModal();
-      fetchData();
+     const token = JSON.parse(localStorage.getItem("access_token"));
+     if (
+       token &&
+       token.expirationDate &&
+       new Date() > new Date(token.expirationDate)
+     ) {
+       // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+       alert("Token is expired, please login again.");
+     } else {
+       // Token còn hiệu lực, tiếp tục sử dụng
+       await axios.put(
+         `http://localhost:5000/api/blog/${id}`,
+         {
+           title,
+           category,
+           description,
+         },
+         {
+           headers: {
+             "Content-Type": "application/json",
+             Authorization: `Bearer ${token}`,
+           },
+         }
+       );
+       handleCloseModal();
+       fetchData();
+     }
     } catch (error) {
       throw new Error(error);
     }
@@ -102,16 +142,30 @@ const Bloglist = () => {
 
   const handleUploadImage = async (id, e) => {
     try {
-      const token = localStorage.getItem("access_token");
-      const formData = new FormData();
-      formData.append("images", e.target.files[0]);
-      await axios.put(`http://localhost:5000/api/blog/upload/${id}`, formData, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "multipart/form-data", // Đặt header "Content-Type" là "multipart/form-data" để gửi dữ liệu dạng form-data
-        },
-      });
-      fetchData();
+      const token = JSON.parse(localStorage.getItem("access_token"));
+      if (
+        token &&
+        token.expirationDate &&
+        new Date() > new Date(token.expirationDate)
+      ) {
+        // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+        alert("Token is expired, please login again.");
+      } else {
+        // Token còn hiệu lực, tiếp tục sử dụng
+         const formData = new FormData();
+         formData.append("images", e.target.files[0]);
+         await axios.put(
+           `http://localhost:5000/api/blog/upload/${id}`,
+           formData,
+           {
+             headers: {
+               Authorization: `Bearer ${token}`,
+               "Content-Type": "multipart/form-data", // Đặt header "Content-Type" là "multipart/form-data" để gửi dữ liệu dạng form-data
+             },
+           }
+         );
+         fetchData();
+      }
     } catch (error) {
       throw new Error(error);
     }
