+++ src/pages/Categorylist.jsx
@@ -13,16 +13,26 @@ const Categorylist = () => {
   });
   const fetchData = async () => {
     try {
-      const token = localStorage.getItem("access_token");
-      const response = await axios.get(
-        "http://localhost:5000/api/prodcategory/",
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      setData(response.data);
+      const token = JSON.parse(localStorage.getItem("access_token"));
+      if (
+        token &&
+        token.expirationDate &&
+        new Date() > new Date(token.expirationDate)
+      ) {
+        // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+        alert("Token is expired, please login again.");
+      } else {
+        // Token còn hiệu lực, tiếp tục sử dụng
+        const response = await axios.get(
+          "http://localhost:5000/api/prodcategory/",
+          {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          }
+        );
+        setData(response.data);
+      }
     } catch (error) {
       throw new Error(error);
     }
@@ -33,13 +43,23 @@ const Categorylist = () => {
   const onDeleteProdCategory = async (id, e) => {
     if (window.confirm("Bạn có chắc chắn muốn xóa dữ liệu này không?")) {
       try {
-        const token = localStorage.getItem("access_token");
-        await axios.delete(`http://localhost:5000/api/prodcategory/${id}`, {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        fetchData();
+        const token = JSON.parse(localStorage.getItem("access_token"));
+        if (
+          token &&
+          token.expirationDate &&
+          new Date() > new Date(token.expirationDate)
+        ) {
+          // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+          alert("Token is expired, please login again.");
+        } else {
+          // Token còn hiệu lực, tiếp tục sử dụng
+         await axios.delete(`http://localhost:5000/api/prodcategory/${id}`, {
+           headers: {
+             Authorization: `Bearer ${token}`,
+           },
+         });
+         fetchData();
+        }
       } catch (error) {
         throw new Error(error);
       }
@@ -55,18 +75,28 @@ const Categorylist = () => {
   const handleUpdateProdCategory = async (e) => {
     const { id, title } = updateData;
     try {
-      const token = localStorage.getItem("access_token");
-      await axios.put(
-        `http://localhost:5000/api/prodcategory/${id}`,
-        { title },
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      handleCloseModal();
-      fetchData();
+      const token = JSON.parse(localStorage.getItem("access_token"));
+      if (
+        token &&
+        token.expirationDate &&
+        new Date() > new Date(token.expirationDate)
+      ) {
+        // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+        alert("Token is expired, please login again.");
+      } else {
+        // Token còn hiệu lực, tiếp tục sử dụng
+          await axios.put(
+            `http://localhost:5000/api/prodcategory/${id}`,
+            { title },
+            {
+              headers: {
+                Authorization: `Bearer ${token}`,
+              },
+            }
+          );
+          handleCloseModal();
+          fetchData();
+      }
     } catch (error) {
       throw new Error(error);
     }
