+++ src/pages/Productlist.jsx
@@ -12,16 +12,27 @@ const Productlist = () => {
   useEffect(() => {
     const fetchCategories = async () => {
       try {
-        const token = localStorage.getItem("access_token");
-        const response = await axios.get(
-          "http://localhost:5000/api/prodcategory/",
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          }
-        );
-        setCategories(response.data);
+        const token = JSON.parse(localStorage.getItem("access_token"));
+        if (
+          token &&
+          token.expirationDate &&
+          new Date() > new Date(token.expirationDate)
+        ) {
+          // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+          alert("Token is expired, please login again.");
+        } else {
+          // Token còn hiệu lực, tiếp tục sử dụng
+          const response = await axios.get(
+            "http://localhost:5000/api/prodcategory/",
+            {
+              headers: {
+                Authorization: `Bearer ${token}`,
+              },
+            }
+          );
+          setCategories(response.data);
+        }
+        
       } catch (error) {
         console.error(error);
       }
@@ -31,13 +42,24 @@ const Productlist = () => {
   useEffect(() => {
     const fetchBrands = async () => {
       try {
-      const token = localStorage.getItem("access_token");
-        const response = await axios.get("http://localhost:5000/api/brand/", {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        setBrands(response.data);
+        const token = JSON.parse(localStorage.getItem("access_token"));
+        if (
+          token &&
+          token.expirationDate &&
+          new Date() > new Date(token.expirationDate)
+        ) {
+          // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+          alert("Token is expired, please login again.");
+        } else {
+          // Token còn hiệu lực, tiếp tục sử dụng
+          const response = await axios.get("http://localhost:5000/api/brand/", {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          });
+          setBrands(response.data);
+        }
+        
       } catch (error) {
         console.error(error);
       }
@@ -56,13 +78,24 @@ const Productlist = () => {
   });
   const fetchData = async () => {
     try {
-      const token = localStorage.getItem("access_token");
-      const response = await axios.get("http://localhost:5000/api/product/", {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      setData(response.data);
+      const token = JSON.parse(localStorage.getItem("access_token"));
+      if (
+        token &&
+        token.expirationDate &&
+        new Date() > new Date(token.expirationDate)
+      ) {
+        // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+        alert("Token is expired, please login again.");
+      } else {
+        // Token còn hiệu lực, tiếp tục sử dụng
+       const response = await axios.get("http://localhost:5000/api/product/", {
+         headers: {
+           Authorization: `Bearer ${token}`,
+         },
+       });
+       setData(response.data);
+      }
+      
     } catch (error) {
       throw new Error(error);
     }
@@ -73,13 +106,24 @@ const Productlist = () => {
   const onDeleteProduct = async (id, e) => {
     if (window.confirm("Bạn có chắc chắn muốn xóa dữ liệu này không?")) {
       try {
-        const token = localStorage.getItem("access_token");
-        await axios.delete(`http://localhost:5000/api/product/${id}`, {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        fetchData();
+        const token = JSON.parse(localStorage.getItem("access_token"));
+        if (
+          token &&
+          token.expirationDate &&
+          new Date() > new Date(token.expirationDate)
+        ) {
+          // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+          alert("Token is expired, please login again.");
+        } else {
+          // Token còn hiệu lực, tiếp tục sử dụng
+          await axios.delete(`http://localhost:5000/api/product/${id}`, {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          });
+          fetchData();
+        }
+        
       } catch (error) {
         throw new Error(error);
       }
@@ -111,26 +155,37 @@ const Productlist = () => {
     const { id, title, description, price, category, brand, quantity, color } =
       updateData;
     try {
-      const token = localStorage.getItem("access_token");
-      await axios.put(
-        `http://localhost:5000/api/product/${id}`,
-        {
-          title,
-          description,
-          price,
-          category,
-          brand,
-          quantity,
-          color,
-        },
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
+      const token = JSON.parse(localStorage.getItem("access_token"));
+      if (
+        token &&
+        token.expirationDate &&
+        new Date() > new Date(token.expirationDate)
+      ) {
+        // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+        alert("Token is expired, please login again.");
+      } else {
+        // Token còn hiệu lực, tiếp tục sử dụng
+        await axios.put(
+          `http://localhost:5000/api/product/${id}`,
+          {
+            title,
+            description,
+            price,
+            category,
+            brand,
+            quantity,
+            color,
           },
-        }
-      );
-      handleCloseModal();
-      fetchData();
+          {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          }
+        );
+        handleCloseModal();
+        fetchData();
+      }
+      
     } catch (error) {
       throw new Error(error);
     }
@@ -139,16 +194,30 @@ const Productlist = () => {
 
   const handleUploadImage = async (id, e) => {
     try {
-      const token = localStorage.getItem("access_token");
-      const formData = new FormData();
-      formData.append("images", e.target.files[0]);
-      await axios.put(`http://localhost:5000/api/product/upload/${id}`, formData, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "multipart/form-data", // Đặt header "Content-Type" là "multipart/form-data" để gửi dữ liệu dạng form-data
-        },
-      });
-      fetchData();
+      const token = JSON.parse(localStorage.getItem("access_token"));
+      if (
+        token &&
+        token.expirationDate &&
+        new Date() > new Date(token.expirationDate)
+      ) {
+        // Token đã hết hạn, xử lý tương ứng (ví dụ: đăng nhập lại)
+        alert("Token is expired, please login again.");
+      } else {
+        // Token còn hiệu lực, tiếp tục sử dụng
+        const formData = new FormData();
+        formData.append("images", e.target.files[0]);
+        await axios.put(
+          `http://localhost:5000/api/product/upload/${id}`,
+          formData,
+          {
+            headers: {
+              Authorization: `Bearer ${token}`,
+              "Content-Type": "multipart/form-data", // Đặt header "Content-Type" là "multipart/form-data" để gửi dữ liệu dạng form-data
+            },
+          }
+        );
+        fetchData();
+      }
     } catch (error) {
       throw new Error(error);
     }
